audio with image switching test
little demo of playing audio, checking for input at the end of the clip, and changing the image according to what came before. basically plug and plop social story images and add the code to change the audio clips (which would be the same as for the images but only with audio sources.) you can install it on your android right now and test this demo if you like!